Move AstroSaveErrors to errors module
